% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bed2Gene_funcs.R
\name{bed2gene}
\alias{bed2gene}
\title{Annotate BED file with gene names}
\usage{
bed2gene(
  file,
  genes = c(),
  geneFile = "",
  prefix = "",
  outDir = pwd(),
  build = "hg19",
  rmChrM = TRUE
)
}
\arguments{
\item{file}{BED file}

\item{genes}{Only output and annotation for intervals that overlap genes in list. Default is an empty vector which results in all genes being output. (optional)}

\item{geneFile}{File containing genes. Only output and annotation for intervals that overlap genes in the file will be output. Alternately, genes can be entered using the gene argument. (optional)}

\item{prefix}{Prefix for output files (optional)}

\item{outDir}{Output directory. If not given results will be printed to pwd(). (optional)}

\item{build}{Build reference for the position coordingates int he BED file (hg19 or hg38)}

\item{rmChrM}{Indicates if mitochondrial genes should be output and annotated (TRUE/FALSE)}
}
\value{
Null
}
\description{
Annotate BED file with gene names based on overlap between the BED interval and coding exons
}
\examples{
bed2gene(file = "MyBedFile.bed", prefilx = "myGenes.bed", rmChrM=TRUE)

}
\seealso{
\code{\link{gene2bed}}
}
\author{
Andrew Skol, \email{askol@luriechildrens.org}
}
\keyword{utilities}
