% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gene2Bed_funcs.R
\name{gene2bed}
\alias{gene2bed}
\title{Create a BED file containing the coding regions of genes}
\usage{
gene2bed(genes, geneFile, prefix, outDir, pad = 0, rmChrM = TRUE)
}
\arguments{
\item{genes}{List of genes for which coding exons should be output coding exons.}

\item{prefix}{Prefix for output files (optional)}

\item{outDir}{Output directory (optional)}

\item{pad}{Amount of padding up and downstream of coding exon (in bp) (default = 0)}

\item{rmChrM}{Indicates if mitochondrial genes should be output (TRUE/FALSE) (default is TRUE)}

\item{genefile}{A file that contains one gene per line}
}
\value{
Null
}
\description{
Create a BED file of coding exons for a user-specified set of genes. Padding can be added#'
}
\examples{
gene2bed(c("TP53","HNF4A"), prefix="MyGenes", pad=10, rmChrM=TRUE)

}
\seealso{
\code{\link{bed2gene}}
}
\author{
Andrew Skol, \email{askol@luriechildrens.org}
}
\keyword{utilities}
